requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.1
django-cors-headers==2.1.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.1
djangorestframework-simplejwt==3.1
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
urllib3==1.22
